Refactor app structure with route-based pages and config
