+++ functions/index.ts
@@ -1,3 +1,3 @@
-export { debitCredits } from "./debitCredits";
+export { debitCredits } from "./src/debitCredits";
 export { stripeWebhook } from "./stripeWebhook";
 export { createCheckoutSession } from "./createCheckoutSession";
