+++ src/lib/api/createLiveStream.js
@@ -6,7 +6,6 @@ export async function createLiveStream() {
     return res.data;
   } catch (err) {
     console.error(err);
-    // mock data in case of error
     return { rtmpUrl: "rtmp://example.com/live", streamKey: "sampleKey", id: "mockId" };
   }
 }
